cli/CLI/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
requests==2.31.0
tabulate==0.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
matplotlib==3.7.1
networkx==3.1
python-dotenv==1.0.0

# Dependencias para el CLI del Orquestador
click==8.1.3
pyyaml==6.0

# Autenticación y HTTP
urllib3==2.1.0

# Para interfaces mejoradas (opcional)
rich==13.7.0
colorama==0.4.6